refactor: tweaks per code review
