+++ packages/inngest/src/experimental.ts
@@ -1,11 +1,7 @@
 // AsyncLocalStorage
 
 export type { AsyncContext } from "./components/execution/als.ts";
-export {
-  getAsyncCtx,
-  getAsyncCtxSync,
-  getAsyncLocalStorage,
-} from "./components/execution/als.ts";
+export { getAsyncCtx } from "./components/execution/als.ts";
 // Extended Traces (OpenTelemetry)
 export type { ExtendedTracesMiddlewareOptions } from "./components/execution/otel/middleware.ts";
 export { extendedTracesMiddleware } from "./components/execution/otel/middleware.ts";
